requirements.txt
aniso8601==10.0.0
Authlib==1.4.1
bcrypt==4.2.1
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
mysql-connector-python==9.2.0
packaging==24.2
pillow==11.1.0
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
